backend/requirements.txt
absl-py==2.2.1
asgiref==3.8.1
astunparse==1.6.3
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
filelock==3.18.0
flake8==7.2.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
isort==6.0.1
joblib==1.4.2
keras==3.9.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mypy_extensions==1.1.0
namex==0.0.8
nodeenv==1.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pickle5==0.0.11
pip-autoremove==0.10.0
platformdirs==4.3.7
pre_commit==4.2.0
protobuf==5.29.4
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sklearn==0.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.30.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
